Harden xperiment discovery and move model utils
Launchers for M1–M5 now only list `xparameters/*.py` modules that define `XPARAMETERS`, preventing helper modules from appearing as broken menu entries. Instrumentation modules were moved into each model’s new `utils/` package (with import paths updated) to separate support code from core simulation files. Docs were updated to codify the `xparameters/utils/` and `<model>/utils/` conventions and to expand PR review guidance with a new maintainer-edits section plus updated section references.
