+++ openwave/xperiments/m2_free_wave/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Supporting scripts for this model: instrumentation, plotting and monitoring.
+
+The model root holds the launcher, the medium and the engines. Anything that only
+observes, records or draws lives here.
+"""
